[postgres] Move gui related code to a separate file
